Layout_TestCase/http/tests/xmlhttprequest/access-control-basic-get-fail-non-simple-expected.txt

CONSOLE MESSAGE: XMLHttpRequest cannot load http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests/xmlhttprequest/resources/access-control-basic-get-fail-non-simple.cgi. Origin http://http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests is not allowed by Access-Control-Allow-Origin.
PASS: Exception thrown. Cross-domain access was denied in 'send'. [NETWORK_ERR: XMLHttpRequest Exception 101].
